pyproject.toml
[project]
name = "expensetracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-cors>=6.0.0",
    "flask-jwt-extended>=4.7.1",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dateutil>=2.9.0.post0",
]
